% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab_functions.R
\name{create_p_ctab}
\alias{create_p_ctab}
\title{Create a color table based on p-values and positive/negative direction of effect}
\usage{
create_p_ctab(x, structure_col, direction_col, pval_col,
  colramp_pos = c("yellow", "red"), colramp_neg = c("cyan", "blue"),
  p_range = c(1e-05, 0.05), default_unknown = TRUE, ...)
}
\description{
Create a color table based on p-values and positive/negative direction of effect
}
