% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{get_vertex_labels}
\alias{get_vertex_labels}
\title{Get vertex labels}
\usage{
get_vertex_labels(object, new_ctab = NULL, labelcol = NULL,
  RGBAcols = NULL, ...)
}
\description{
Simplified version of replace_ctab
}
