% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{replace_ctab}
\alias{replace_ctab}
\title{Insert new color table in annotation_construct}
\usage{
replace_ctab(object, new_ctab, labelcol, RGBAcols, ...)
}
\description{
Insert new color table in annotation_construct
}
