% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_functions.R
\name{write_annotation}
\alias{write_annotation}
\title{Write annotation_construct objects to Freesurfer annotation file}
\usage{
write_annotation(object, file, convert2integer = FALSE, ...)
}
\description{
Write annotation_construct objects to Freesurfer annotation file
}
